sc_server/pyproject.toml
[tool.poetry]
name = "sc-server"
version = "0.1.0"
description = "SlideCaster Server"
authors = ["Dessera <lord.changed@foxmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.5"}
uvicorn = {extras = ["standard"], version = "^0.32.0"}
python-dotenv = "^1.0.1"
typer = "^0.13.0"
mediapipe = "^0.10.18"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
